refactor(home): collapse the duplicate weekday mapping
Extracting ClassTableSelection.dayTimeFrom left two copies of the
Calendar-to-Monday-first conversion, differing only in their `else`
branch — one fell through to Monday, the other to Sunday. Identical for
every real input, which is exactly why it would have stayed that way.

weekdayIndex moves to data.CourseRosterMerge alongside the other rules
both screens share, and its test moves with it. HomeCourseMerge is down
to the one thing that is genuinely Home-specific, and its header no
longer claims to own the submission safety net that moved out.

270 unit tests pass; all three variants compile.
